Relocated CSP to root of GoCSI
This patch relocates the CSP package to the root of the GoCSI project
and updates related references. The script "csp.sh" has been renamed
"gocsi.sh".
